+++ monitor_suite_agent/__init__.py
@@ -1,4 +1,4 @@
 """Monitor Suite Agent."""
 
-__version__ = "2.5.1"
+__version__ = "2.8.0"
 __copyright__ = "Copyright (c) 2026 Toast"